requirements.txt
# Python 3.11+ required
streamlit>=1.31
pandas>=2.1
openpyxl>=3.1
PyYAML>=6.0
msal>=1.26
requests>=2.31
pytest>=8.0
